Separate board specific thermometer component. With tests
