+++ apps/src/lib/kits/maker/boards/circuitPlayground/PlaygroundComponents.js
@@ -14,7 +14,7 @@ import _ from 'lodash';
 import five from '@code-dot-org/johnny-five';
 import PlaygroundIO from 'playground-io';
 import PlaygroundButton from './Button';
-import PlaygroundThermometer from '../../Thermometer';
+import PlaygroundThermometer from './Thermometer';
 import TouchSensor from './TouchSensor';
 import Piezo from './Piezo';
 import NeoPixel from './NeoPixel';
