+++ apps/src/lib/kits/maker/boards/microBit/MicroBitComponents.js
@@ -4,7 +4,7 @@
 import MicroBitButton from './MicroBitButton';
 import LedMatrix from './LedMatrix';
 import Accelerometer from './Accelerometer';
-import {MicroBitThermometer} from '../../Thermometer';
+import MicroBitThermometer from './MicroBitThermometer';
 
 /**
  * Initializes a set of components for the currently
